chore: Implement cleaning prompt and dry run option for directory cleaning

+++ cli/src/main.rs
@@ -19,6 +19,9 @@ struct Args {
     depth: u16,
 
     path: Option<PathBuf>,
+
+    #[arg(long, help = "Dry Run", default_value = "false")]
+    dry_run: bool,
 }
 
 #[derive(Subcommand, Clone, Debug)]
@@ -31,30 +34,52 @@ enum Commands {
     },
 }
 
-fn ask_clean(target: &AnalyzeTarget) -> Result<()> {
-    let parent = target.path.parent().unwrap();
-    let dir_name: String = target.path.file_name().unwrap().to_string_lossy().into();
-    let size = get_size(target.path.clone())?;
-    println!(
-        "{}\n\t└─ {} ({})",
-        parent.display(),
-        dir_name,
-        human_bytes(size as f64)
-    );
-    print!("Do you want to clean this directory? [y/N]:");
-    std::io::stdout().flush()?;
-    let mut input = String::new();
-    std::io::stdin().read_line(&mut input)?;
-    if input.trim().to_lowercase() == "y" {
-        println!("Cleaning {}", target.path.display());
-        std::fs::remove_dir_all(&target.path)?;
+struct Cleaner {
+    bytes_cleaned: u128,
+    dry_run: bool,
+}
+impl Cleaner {
+    fn new(dry_run: bool) -> Self {
+        Cleaner {
+            bytes_cleaned: 0,
+            dry_run,
+        }
     }
-    Ok(())
+    fn prompt_clean(&mut self, target: &AnalyzeTarget) -> Result<()> {
+        let parent = target.path.parent().unwrap();
+        let dir_name: String = target.path.file_name().unwrap().to_string_lossy().into();
+        let size = get_size(target.path.clone())?;
+        self.bytes_cleaned += size as u128;
+        println!(
+            "{}\n  └─ {} ({})",
+            parent.display(),
+            dir_name,
+            human_bytes(size as f64)
+        );
+        if self.dry_run {
+            print!("(Dry Run)  ");
+        }
+        print!("Do you want to clean this directory? [y/N]:");
+        std::io::stdout().flush()?;
+        let mut input = String::new();
+        std::io::stdin().read_line(&mut input)?;
+        if input.trim().to_lowercase() == "y" {
+            if self.dry_run {
+                print!("(Dry Run)  ");
+            }
+            println!("Cleaning {}", target.path.display());
+            if !self.dry_run {
+                std::fs::remove_dir_all(&target.path)?;
+            }
+        }
+        Ok(())
+    }
 }
 
 fn main() -> Result<()> {
     color_eyre::install()?;
     let args = Args::parse();
+
     match args.command {
         Some(Commands::FindDirtyGit { path, depth }) => {
             let path = path.unwrap_or_else(|| PathBuf::from("."));
@@ -74,10 +99,15 @@ fn main() -> Result<()> {
             let mut removable_scanner = get_project_garbage_scanner(path.as_path(), args.depth);
             removable_scanner.scan();
             let mut targets = Vec::new();
+            let mut cleaner = Cleaner::new(args.dry_run);
             while let Ok(target) = removable_scanner.task_rx.recv() {
-                ask_clean(&target);
+                cleaner.prompt_clean(&target)?;
                 targets.push(target);
             }
+            println!(
+                "Total Bytes Cleaned: {}",
+                human_bytes(cleaner.bytes_cleaned as f64)
+            );
             // AnalyzeTargets(targets).to_table().printstd();
         }
     }
